requirements.txt
pandas
python-dotenv
streamlit
openai
pydantic
pyyaml
fastapi
uvicorn
PyPDF2